Algumas atualizacoes no sistema.py e renomeado main.py para jogo.py
