+++ prototipo/sistema.py
@@ -1,0 +1,3 @@
+import jogo
+
+jogo.loop_principal()
